AOC 2017 - 14 Solution
This one was... a mixed bag.  I had trouble on part one because I kind of forgot which parts of the knot hash from day 10 I needed. The next part was not too bad, but it took me awhile (after writing pseudocode) to realize I was forgetting to check if the variations were already in the region, which resulted in the .pop() returning -1 and wrapping around the array (oh python and your quirks).