AOC 2017 - 16 Solution
The first part of this "dance" was pretty straight forward. Just some simple array manipulations. The second part was a great challenge. I tried to brute force it, but after waiting 20 minutes I realized I needed a better solution to handle ONE BILLION dances. I figured that eventually, the orientation of the programs must cycle through older iterations in a sort of loop. I then found this cycle and took the modulus of the amount of times you dance to find where in the cycle you would end up. One of my favorite challenges so far!