AOC 2017 - 18 Solution
Jeez. I attempted to mess with eval/exec for the first part and got confused quickly. However, it turned out okay and worked. But man, part two killed me. I spent so long trying to work out the logic errors in my program, only for the bugs to be dumb errors (jgz can have a int input, .isdigit() returns false for negative numbers [this one hurt the most]). But hey, it was interesting to look at assembly-like code.