AOC 2015 - 7 Solution
That was a bit difficult. The parsing was not very friendly, not to mention the whole signed binary numbers in python led to some python documentation reading. Also part 2 was EXTREMELY vague on what i meant by "overriding". I assumed it meant start with that signal, but eventually figured out it meant keep that signal as a constant!

AOC 2015 - 7 Merge