AOC 2015 - 15 Solution
Whew. Took me a bit to come up with a solution that spread 100 out among the ingredients. Decided to go with for loops and only take arrays with a sum of 100. After some bug fixing (did not see the "return 0 if a SINGLE total is negative" part, that cost me some time...) and optimization, I got a pretty fast solution that I am semi-proud of.