AOC 2016 - 8 Solution
Man, that probably has to be my favorite day ever in AOC. The first part was a moderate task that wasn't too hard, but very fun to implement. The second part was pure genius, and I appreciate puzzles like this where the output isn't a string literal, but instead a picture.  Best one yet!