AOC 2021 - 7 Solution
Would have made it on the leaderboard if it wasn't for grabbing the wrong input file... sad. Easy day that you can just brute force. Might take some time to see if there is a way I can optimize it somehow! (Also, shoutout to the crabs for their hard work!)