AOC 2021 - 8 Solution
Part one was super easy (just a lot of convoluted readings to get through). Part two... shook me. It took me forever to figure out what it was even saying in the first place, and once I did, I was lost on how to solve it. Thats when my friend brute-force came in, and he works perfectly!