AOC 2021 - 8 Optimization
Switched from brute force to a more systematic and logical way of thinking.  I made a dictionary of values of the signals compared to the "1" signal and "4" signal along with the length of the signal. This generates a unique tuple for each number that isn't easy. Fun set algebra concepts!